+++ plugin/src/test/groovy/org/openbakery/packaging/PackageTask_WatchAppSpecification.groovy
@@ -128,6 +128,10 @@ class PackageTask_WatchAppSpecification extends Specification {
 
 	}
 
+	void includeLibswiftRemoteMirrorIntoExampleApp() {
+		FileUtils.writeStringToFile(new File(applicationBundle, "libswiftRemoteMirror.dylib"), "dummy");
+		FileUtils.writeStringToFile(new File(watchkitExtensionBundle, "libswiftRemoteMirror.dylib"), "dummy");
+	}
 
 	List<String> codesignLibCommand(String path) {
 		File payloadApp = new File(packageTask.outputPath, path)
@@ -216,4 +220,29 @@ class PackageTask_WatchAppSpecification extends Specification {
 
 	}
 
+	def "remove libswiftRemoteMirror.dylib from app"() {
+		given:
+		createExampleApp()
+		includeLibswiftRemoteMirrorIntoExampleApp()
+
+		when:
+		packageTask.packageApplication()
+
+		then:
+		!(new File(packageTask.outputPath, "Payload/ExampleWatchKit.app/libswiftRemoteMirror.dylib").exists())
+
+	}
+
+	def "remove libswiftRemoteMirror.dylib from app extenstion"() {
+		given:
+		createExampleApp()
+		includeLibswiftRemoteMirrorIntoExampleApp()
+
+		when:
+		packageTask.packageApplication()
+
+		then:
+		!(new File(packageTask.outputPath, "Payload/ExampleWatchKit.app/Watch/ExampleWatchkit WatchKit App.app/PlugIns/ExampleWatchkit WatchKit Extension.appex/libswiftRemoteMirror.dylib").exists())
+
+	}
 }